+++ src/test/java/picocli/CommandLineArityTest.java
@@ -505,7 +505,7 @@ public class CommandLineArityTest {
             CommandLine.populateCommand(new BooleanOptionsArity0_nAndParameters(), "-rv234 -bool".split(" "));
             fail("Expected exception");
         } catch (UnmatchedArgumentException ok) {
-            assertEquals("Unmatched argument [-234]", ok.getMessage());
+            assertEquals("Unknown option: -234", ok.getMessage());
         }
     }
     @Test
@@ -851,13 +851,13 @@ public class CommandLineArityTest {
             CommandLine.populateCommand(new NonVarArgArrayParamsZeroArity(), "a", "b", "c");
             fail("Expected UnmatchedArgumentException");
         } catch (UnmatchedArgumentException ex) {
-            assertEquals("Unmatched arguments [a, b, c]", ex.getMessage());
+            assertEquals("Unmatched arguments: a, b, c", ex.getMessage());
         }
         try {
             CommandLine.populateCommand(new NonVarArgArrayParamsZeroArity(), "a");
             fail("Expected UnmatchedArgumentException");
         } catch (UnmatchedArgumentException ex) {
-            assertEquals("Unmatched argument [a]", ex.getMessage());
+            assertEquals("Unmatched argument: a", ex.getMessage());
         }
         NonVarArgArrayParamsZeroArity params = CommandLine.populateCommand(new NonVarArgArrayParamsZeroArity());
         assertEquals(null, params.params);