+++ src/test/java/picocli/LenientParsingTest.java
@@ -74,7 +74,7 @@ public class LenientParsingTest {
         CommandLine commandLine = new CommandLine(spec);
         commandLine.parse("-c", "1", "2", "3");
         assertEquals(1, commandLine.getParseResult().errors().size());
-        assertEquals("Unmatched arguments [2, 3]", commandLine.getParseResult().errors().get(0).getMessage());
+        assertEquals("Unmatched arguments: 2, 3", commandLine.getParseResult().errors().get(0).getMessage());
     }
 
     @Test
@@ -88,7 +88,7 @@ public class LenientParsingTest {
         CommandLine commandLine = new CommandLine(spec);
         commandLine.parse("1", "2", "3");
         assertEquals(1, commandLine.getParseResult().errors().size());
-        assertEquals("Unmatched arguments [2, 3]", commandLine.getParseResult().errors().get(0).getMessage());
+        assertEquals("Unmatched arguments: 2, 3", commandLine.getParseResult().errors().get(0).getMessage());
     }
     @Test
     public void testMissingRequiredParams() {
@@ -182,11 +182,11 @@ public class LenientParsingTest {
         cmd.getCommandSpec().parser().collectErrors(true);
         cmd.parse("a", "b", "c");
         assertEquals(1, cmd.getParseResult().errors().size());
-        assertEquals("Unmatched arguments [a, b, c]", cmd.getParseResult().errors().get(0).getMessage());
+        assertEquals("Unmatched arguments: a, b, c", cmd.getParseResult().errors().get(0).getMessage());
 
         cmd.parse("a");
         assertEquals(1, cmd.getParseResult().errors().size());
-        assertEquals("Unmatched argument [a]", cmd.getParseResult().errors().get(0).getMessage());
+        assertEquals("Unmatched argument: a", cmd.getParseResult().errors().get(0).getMessage());
     }
     @Test
     public void testSingleValueFieldDefaultMinArityIsOne() {
@@ -215,7 +215,7 @@ public class LenientParsingTest {
         cmd.getCommandSpec().parser().collectErrors(true);
         cmd.parse("-rv234 -bool".split(" "));
         assertEquals(1, cmd.getParseResult().errors().size());
-        assertEquals("Unmatched argument [-234]", cmd.getParseResult().errors().get(0).getMessage());
+        assertEquals("Unknown option: -234", cmd.getParseResult().errors().get(0).getMessage());
     }
 
     @Test
@@ -228,7 +228,7 @@ public class LenientParsingTest {
         }
         assertMissing(Arrays.asList("Expected parameter 2 (of 2 mandatory parameters) for option '-a' but found '-a'",
                 "option '-a' at index 0 (<a>) requires at least 2 values, but only 1 were specified: [2]",
-                "Unmatched argument [2]"),
+                "Unmatched argument: 2"),
                 new Cmd(), "-a", "1", "-a", "2");
 
         assertMissing(Arrays.asList("Expected parameter 2 (of 2 mandatory parameters) for option '-a' but found '-v'"),
@@ -260,7 +260,7 @@ public class LenientParsingTest {
         }
         assertMissing(Arrays.asList("Expected parameter 2 (of 2 mandatory parameters) for option '-a' but found '-a'",
                 "option '-a' at index 0 (<String=String>) requires at least 2 values, but only 1 were specified: [C=D]",
-                "Unmatched argument [C=D]"),
+                "Unmatched argument: C=D"),
                 new Cmd(), "-a", "A=B", "-a", "C=D");
 
         assertMissing(Arrays.asList("Expected parameter 2 (of 2 mandatory parameters) for option '-a' but found '-v'"),
@@ -304,7 +304,7 @@ public class LenientParsingTest {
         assertEquals(prefix, ex.getMessage().substring(0, prefix.length()));
         assertEquals(suffix, ex.getMessage().substring(ex.getMessage().length() - suffix.length(), ex.getMessage().length()));
 
-        assertEquals("Unmatched argument [c]", cmd.getParseResult().errors().get(1).getMessage());
+        assertEquals("Unmatched argument: c", cmd.getParseResult().errors().get(1).getMessage());
     }
     @Test
     public void testTimeFormatHHmmssSSSInvalidError() throws ParseException {